requirements.txt
discord.py
feedparser

python-dotenv
aiohttp
colorama
asyncpg
